Add: Migrations and Removing mistakes

+++ RestApiDDD.Application/DTOs/ProductDto.cs
@@ -3,7 +3,7 @@
     public class ProductDto
     {
         public Guid Id { get; set; }
-        public string Name { get; set; }    
-        public decimal Price { get; set; }  
+        public string Name { get; set; }
+        public double Price { get; set; }  
     }
 }
